enhance create event api

+++ src/main/java/com/eventra/dto/TicketResponse.java
@@ -2,9 +2,6 @@ package com.eventra.dto;
 
 import lombok.Data;
 
-import com.eventra.dto.EventResponse;
-
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -13,7 +10,7 @@ public class TicketResponse {
     private UUID id;
     private EventResponse event;
     private String ticketCategory;
-    private BigDecimal price;
+    private Integer price;
     private Integer quota;
     private LocalDateTime createdAt;
     private String createdBy;
